third_party/WebKit/LayoutTests/css-parser/color3_hsl-expected.txt

CONSOLE WARNING: line 3: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Color test for resources/css-parsing-tests/color3_hsl.json

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorTest.parseColor("hsl(0, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 100%, 0%)") is "rgb(0, 0, 0)"
FAIL colorTest.parseColor("hsl(0, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(30, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(60, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(90, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(120, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(150, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(180, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(210, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(240, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(270, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(300, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(330, 0%, 12.5%)") should be rgb(31, 31, 31). Was rgb(32, 32, 32).
FAIL colorTest.parseColor("hsl(0, 12.5%, 12.5%)") should be rgb(35, 27, 27). Was rgb(36, 28, 28).
FAIL colorTest.parseColor("hsl(30, 12.5%, 12.5%)") should be rgb(35, 31, 27). Was rgb(36, 32, 28).
FAIL colorTest.parseColor("hsl(60, 12.5%, 12.5%)") should be rgb(35, 35, 27). Was rgb(36, 36, 28).
FAIL colorTest.parseColor("hsl(90, 12.5%, 12.5%)") should be rgb(31, 35, 27). Was rgb(32, 36, 28).
FAIL colorTest.parseColor("hsl(120, 12.5%, 12.5%)") should be rgb(27, 35, 27). Was rgb(28, 36, 28).
FAIL colorTest.parseColor("hsl(150, 12.5%, 12.5%)") should be rgb(27, 35, 31). Was rgb(28, 36, 32).
FAIL colorTest.parseColor("hsl(180, 12.5%, 12.5%)") should be rgb(27, 35, 35). Was rgb(28, 36, 36).
FAIL colorTest.parseColor("hsl(210, 12.5%, 12.5%)") should be rgb(27, 31, 35). Was rgb(28, 32, 36).
FAIL colorTest.parseColor("hsl(240, 12.5%, 12.5%)") should be rgb(27, 27, 35). Was rgb(28, 28, 36).
FAIL colorTest.parseColor("hsl(270, 12.5%, 12.5%)") should be rgb(31, 27, 35). Was rgb(32, 28, 36).
FAIL colorTest.parseColor("hsl(300, 12.5%, 12.5%)") should be rgb(35, 27, 35). Was rgb(36, 28, 36).
FAIL colorTest.parseColor("hsl(330, 12.5%, 12.5%)") should be rgb(35, 27, 31). Was rgb(36, 28, 32).
FAIL colorTest.parseColor("hsl(0, 25%, 12.5%)") should be rgb(39, 23, 23). Was rgb(40, 24, 24).
FAIL colorTest.parseColor("hsl(30, 25%, 12.5%)") should be rgb(39, 31, 23). Was rgb(40, 32, 24).
FAIL colorTest.parseColor("hsl(60, 25%, 12.5%)") should be rgb(39, 39, 23). Was rgb(40, 40, 24).
FAIL colorTest.parseColor("hsl(90, 25%, 12.5%)") should be rgb(31, 39, 23). Was rgb(32, 40, 24).
FAIL colorTest.parseColor("hsl(120, 25%, 12.5%)") should be rgb(23, 39, 23). Was rgb(24, 40, 24).
FAIL colorTest.parseColor("hsl(150, 25%, 12.5%)") should be rgb(23, 39, 31). Was rgb(24, 40, 32).
FAIL colorTest.parseColor("hsl(180, 25%, 12.5%)") should be rgb(23, 39, 39). Was rgb(24, 40, 40).
FAIL colorTest.parseColor("hsl(210, 25%, 12.5%)") should be rgb(23, 31, 39). Was rgb(24, 32, 40).
FAIL colorTest.parseColor("hsl(240, 25%, 12.5%)") should be rgb(23, 23, 39). Was rgb(24, 24, 40).
FAIL colorTest.parseColor("hsl(270, 25%, 12.5%)") should be rgb(31, 23, 39). Was rgb(32, 24, 40).
FAIL colorTest.parseColor("hsl(300, 25%, 12.5%)") should be rgb(39, 23, 39). Was rgb(40, 24, 40).
FAIL colorTest.parseColor("hsl(330, 25%, 12.5%)") should be rgb(39, 23, 31). Was rgb(40, 24, 32).
FAIL colorTest.parseColor("hsl(0, 37.5%, 12.5%)") should be rgb(43, 19, 19). Was rgb(44, 20, 20).
FAIL colorTest.parseColor("hsl(30, 37.5%, 12.5%)") should be rgb(43, 31, 19). Was rgb(44, 32, 20).
FAIL colorTest.parseColor("hsl(60, 37.5%, 12.5%)") should be rgb(43, 43, 19). Was rgb(44, 44, 20).
FAIL colorTest.parseColor("hsl(90, 37.5%, 12.5%)") should be rgb(31, 43, 19). Was rgb(32, 44, 20).
FAIL colorTest.parseColor("hsl(120, 37.5%, 12.5%)") should be rgb(19, 43, 19). Was rgb(20, 44, 20).
FAIL colorTest.parseColor("hsl(150, 37.5%, 12.5%)") should be rgb(19, 43, 31). Was rgb(20, 44, 32).
FAIL colorTest.parseColor("hsl(180, 37.5%, 12.5%)") should be rgb(19, 43, 43). Was rgb(20, 44, 44).
FAIL colorTest.parseColor("hsl(210, 37.5%, 12.5%)") should be rgb(19, 31, 43). Was rgb(20, 32, 44).
FAIL colorTest.parseColor("hsl(240, 37.5%, 12.5%)") should be rgb(19, 19, 43). Was rgb(20, 20, 44).
FAIL colorTest.parseColor("hsl(270, 37.5%, 12.5%)") should be rgb(31, 19, 43). Was rgb(32, 20, 44).
FAIL colorTest.parseColor("hsl(300, 37.5%, 12.5%)") should be rgb(43, 19, 43). Was rgb(44, 20, 44).
FAIL colorTest.parseColor("hsl(330, 37.5%, 12.5%)") should be rgb(43, 19, 31). Was rgb(44, 20, 32).
FAIL colorTest.parseColor("hsl(0, 50%, 12.5%)") should be rgb(47, 15, 15). Was rgb(48, 16, 16).
FAIL colorTest.parseColor("hsl(30, 50%, 12.5%)") should be rgb(47, 31, 15). Was rgb(48, 32, 16).
FAIL colorTest.parseColor("hsl(60, 50%, 12.5%)") should be rgb(47, 47, 15). Was rgb(48, 48, 16).
FAIL colorTest.parseColor("hsl(90, 50%, 12.5%)") should be rgb(31, 47, 15). Was rgb(32, 48, 16).
FAIL colorTest.parseColor("hsl(120, 50%, 12.5%)") should be rgb(15, 47, 15). Was rgb(16, 48, 16).
FAIL colorTest.parseColor("hsl(150, 50%, 12.5%)") should be rgb(15, 47, 31). Was rgb(16, 48, 32).
FAIL colorTest.parseColor("hsl(180, 50%, 12.5%)") should be rgb(15, 47, 47). Was rgb(16, 48, 48).
FAIL colorTest.parseColor("hsl(210, 50%, 12.5%)") should be rgb(15, 31, 47). Was rgb(16, 32, 48).
FAIL colorTest.parseColor("hsl(240, 50%, 12.5%)") should be rgb(15, 15, 47). Was rgb(16, 16, 48).
FAIL colorTest.parseColor("hsl(270, 50%, 12.5%)") should be rgb(31, 15, 47). Was rgb(32, 16, 48).
FAIL colorTest.parseColor("hsl(300, 50%, 12.5%)") should be rgb(47, 15, 47). Was rgb(48, 16, 48).
FAIL colorTest.parseColor("hsl(330, 50%, 12.5%)") should be rgb(47, 15, 31). Was rgb(48, 16, 32).
FAIL colorTest.parseColor("hsl(0, 62.5%, 12.5%)") should be rgb(51, 11, 11). Was rgb(52, 12, 12).
FAIL colorTest.parseColor("hsl(30, 62.5%, 12.5%)") should be rgb(51, 31, 11). Was rgb(52, 32, 12).
FAIL colorTest.parseColor("hsl(60, 62.5%, 12.5%)") should be rgb(51, 51, 11). Was rgb(52, 52, 12).
FAIL colorTest.parseColor("hsl(90, 62.5%, 12.5%)") should be rgb(31, 51, 11). Was rgb(32, 52, 12).
FAIL colorTest.parseColor("hsl(120, 62.5%, 12.5%)") should be rgb(11, 51, 11). Was rgb(12, 52, 12).
FAIL colorTest.parseColor("hsl(150, 62.5%, 12.5%)") should be rgb(11, 51, 31). Was rgb(12, 52, 32).
FAIL colorTest.parseColor("hsl(180, 62.5%, 12.5%)") should be rgb(11, 51, 51). Was rgb(12, 52, 52).
FAIL colorTest.parseColor("hsl(210, 62.5%, 12.5%)") should be rgb(11, 31, 51). Was rgb(12, 32, 52).
FAIL colorTest.parseColor("hsl(240, 62.5%, 12.5%)") should be rgb(11, 11, 51). Was rgb(12, 12, 52).
FAIL colorTest.parseColor("hsl(270, 62.5%, 12.5%)") should be rgb(31, 11, 51). Was rgb(32, 12, 52).
FAIL colorTest.parseColor("hsl(300, 62.5%, 12.5%)") should be rgb(51, 11, 51). Was rgb(52, 12, 52).
FAIL colorTest.parseColor("hsl(330, 62.5%, 12.5%)") should be rgb(51, 11, 31). Was rgb(52, 12, 32).
FAIL colorTest.parseColor("hsl(0, 75%, 12.5%)") should be rgb(55, 7, 7). Was rgb(56, 8, 8).
FAIL colorTest.parseColor("hsl(30, 75%, 12.5%)") should be rgb(55, 31, 7). Was rgb(56, 32, 8).
FAIL colorTest.parseColor("hsl(60, 75%, 12.5%)") should be rgb(55, 55, 7). Was rgb(56, 56, 8).
FAIL colorTest.parseColor("hsl(90, 75%, 12.5%)") should be rgb(31, 55, 7). Was rgb(32, 56, 8).
FAIL colorTest.parseColor("hsl(120, 75%, 12.5%)") should be rgb(7, 55, 7). Was rgb(8, 56, 8).
FAIL colorTest.parseColor("hsl(150, 75%, 12.5%)") should be rgb(7, 55, 31). Was rgb(8, 56, 32).
FAIL colorTest.parseColor("hsl(180, 75%, 12.5%)") should be rgb(7, 55, 55). Was rgb(8, 56, 56).
FAIL colorTest.parseColor("hsl(210, 75%, 12.5%)") should be rgb(7, 31, 55). Was rgb(8, 32, 56).
FAIL colorTest.parseColor("hsl(240, 75%, 12.5%)") should be rgb(7, 7, 55). Was rgb(8, 8, 56).
FAIL colorTest.parseColor("hsl(270, 75%, 12.5%)") should be rgb(31, 7, 55). Was rgb(32, 8, 56).
FAIL colorTest.parseColor("hsl(300, 75%, 12.5%)") should be rgb(55, 7, 55). Was rgb(56, 8, 56).
FAIL colorTest.parseColor("hsl(330, 75%, 12.5%)") should be rgb(55, 7, 31). Was rgb(56, 8, 32).
FAIL colorTest.parseColor("hsl(0, 87.5%, 12.5%)") should be rgb(59, 3, 3). Was rgb(60, 4, 4).
FAIL colorTest.parseColor("hsl(30, 87.5%, 12.5%)") should be rgb(59, 31, 3). Was rgb(60, 32, 4).
FAIL colorTest.parseColor("hsl(60, 87.5%, 12.5%)") should be rgb(59, 59, 3). Was rgb(60, 60, 4).
FAIL colorTest.parseColor("hsl(90, 87.5%, 12.5%)") should be rgb(31, 59, 3). Was rgb(32, 60, 4).
FAIL colorTest.parseColor("hsl(120, 87.5%, 12.5%)") should be rgb(3, 59, 3). Was rgb(4, 60, 4).
FAIL colorTest.parseColor("hsl(150, 87.5%, 12.5%)") should be rgb(3, 59, 31). Was rgb(4, 60, 32).
FAIL colorTest.parseColor("hsl(180, 87.5%, 12.5%)") should be rgb(3, 59, 59). Was rgb(4, 60, 60).
FAIL colorTest.parseColor("hsl(210, 87.5%, 12.5%)") should be rgb(3, 31, 59). Was rgb(4, 32, 60).
FAIL colorTest.parseColor("hsl(240, 87.5%, 12.5%)") should be rgb(3, 3, 59). Was rgb(4, 4, 60).
FAIL colorTest.parseColor("hsl(270, 87.5%, 12.5%)") should be rgb(31, 3, 59). Was rgb(32, 4, 60).
FAIL colorTest.parseColor("hsl(300, 87.5%, 12.5%)") should be rgb(59, 3, 59). Was rgb(60, 4, 60).
FAIL colorTest.parseColor("hsl(330, 87.5%, 12.5%)") should be rgb(59, 3, 31). Was rgb(60, 4, 32).
FAIL colorTest.parseColor("hsl(0, 100%, 12.5%)") should be rgb(63, 0, 0). Was rgb(64, 0, 0).
FAIL colorTest.parseColor("hsl(30, 100%, 12.5%)") should be rgb(63, 31, 0). Was rgb(64, 32, 0).
FAIL colorTest.parseColor("hsl(60, 100%, 12.5%)") should be rgb(63, 63, 0). Was rgb(64, 64, 0).
FAIL colorTest.parseColor("hsl(90, 100%, 12.5%)") should be rgb(31, 63, 0). Was rgb(32, 64, 0).
FAIL colorTest.parseColor("hsl(120, 100%, 12.5%)") should be rgb(0, 63, 0). Was rgb(0, 64, 0).
FAIL colorTest.parseColor("hsl(150, 100%, 12.5%)") should be rgb(0, 63, 31). Was rgb(0, 64, 32).
FAIL colorTest.parseColor("hsl(180, 100%, 12.5%)") should be rgb(0, 63, 63). Was rgb(0, 64, 64).
FAIL colorTest.parseColor("hsl(210, 100%, 12.5%)") should be rgb(0, 31, 63). Was rgb(0, 32, 64).
FAIL colorTest.parseColor("hsl(240, 100%, 12.5%)") should be rgb(0, 0, 63). Was rgb(0, 0, 64).
FAIL colorTest.parseColor("hsl(270, 100%, 12.5%)") should be rgb(31, 0, 63). Was rgb(32, 0, 64).
FAIL colorTest.parseColor("hsl(300, 100%, 12.5%)") should be rgb(63, 0, 63). Was rgb(64, 0, 64).
FAIL colorTest.parseColor("hsl(330, 100%, 12.5%)") should be rgb(63, 0, 31). Was rgb(64, 0, 32).
FAIL colorTest.parseColor("hsl(0, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(30, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(60, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(90, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(120, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(150, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(180, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(210, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(240, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(270, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(300, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(330, 0%, 25%)") should be rgb(63, 63, 63). Was rgb(64, 64, 64).
FAIL colorTest.parseColor("hsl(0, 12.5%, 25%)") should be rgb(71, 55, 55). Was rgb(72, 56, 56).
FAIL colorTest.parseColor("hsl(30, 12.5%, 25%)") should be rgb(71, 63, 55). Was rgb(72, 64, 56).
FAIL colorTest.parseColor("hsl(60, 12.5%, 25%)") should be rgb(71, 71, 55). Was rgb(72, 72, 56).
FAIL colorTest.parseColor("hsl(90, 12.5%, 25%)") should be rgb(63, 71, 55). Was rgb(64, 72, 56).
FAIL colorTest.parseColor("hsl(120, 12.5%, 25%)") should be rgb(55, 71, 55). Was rgb(56, 72, 56).
FAIL colorTest.parseColor("hsl(150, 12.5%, 25%)") should be rgb(55, 71, 63). Was rgb(56, 72, 64).
FAIL colorTest.parseColor("hsl(180, 12.5%, 25%)") should be rgb(55, 71, 71). Was rgb(56, 72, 72).
FAIL colorTest.parseColor("hsl(210, 12.5%, 25%)") should be rgb(55, 63, 71). Was rgb(56, 64, 72).
FAIL colorTest.parseColor("hsl(240, 12.5%, 25%)") should be rgb(55, 55, 71). Was rgb(56, 56, 72).
FAIL colorTest.parseColor("hsl(270, 12.5%, 25%)") should be rgb(63, 55, 71). Was rgb(64, 56, 72).
FAIL colorTest.parseColor("hsl(300, 12.5%, 25%)") should be rgb(71, 55, 71). Was rgb(72, 56, 72).
FAIL colorTest.parseColor("hsl(330, 12.5%, 25%)") should be rgb(71, 55, 63). Was rgb(72, 56, 64).
FAIL colorTest.parseColor("hsl(0, 25%, 25%)") should be rgb(79, 47, 47). Was rgb(80, 48, 48).
FAIL colorTest.parseColor("hsl(30, 25%, 25%)") should be rgb(79, 63, 47). Was rgb(80, 64, 48).
FAIL colorTest.parseColor("hsl(60, 25%, 25%)") should be rgb(79, 79, 47). Was rgb(80, 80, 48).
FAIL colorTest.parseColor("hsl(90, 25%, 25%)") should be rgb(63, 79, 47). Was rgb(64, 80, 48).
FAIL colorTest.parseColor("hsl(120, 25%, 25%)") should be rgb(47, 79, 47). Was rgb(48, 80, 48).
FAIL colorTest.parseColor("hsl(150, 25%, 25%)") should be rgb(47, 79, 63). Was rgb(48, 80, 64).
FAIL colorTest.parseColor("hsl(180, 25%, 25%)") should be rgb(47, 79, 79). Was rgb(48, 80, 80).
FAIL colorTest.parseColor("hsl(210, 25%, 25%)") should be rgb(47, 63, 79). Was rgb(48, 64, 80).
FAIL colorTest.parseColor("hsl(240, 25%, 25%)") should be rgb(47, 47, 79). Was rgb(48, 48, 80).
FAIL colorTest.parseColor("hsl(270, 25%, 25%)") should be rgb(63, 47, 79). Was rgb(64, 48, 80).
FAIL colorTest.parseColor("hsl(300, 25%, 25%)") should be rgb(79, 47, 79). Was rgb(80, 48, 80).
FAIL colorTest.parseColor("hsl(330, 25%, 25%)") should be rgb(79, 47, 63). Was rgb(80, 48, 64).
FAIL colorTest.parseColor("hsl(0, 37.5%, 25%)") should be rgb(87, 39, 39). Was rgb(88, 40, 40).
FAIL colorTest.parseColor("hsl(30, 37.5%, 25%)") should be rgb(87, 63, 39). Was rgb(88, 64, 40).
FAIL colorTest.parseColor("hsl(60, 37.5%, 25%)") should be rgb(87, 87, 39). Was rgb(88, 88, 40).
FAIL colorTest.parseColor("hsl(90, 37.5%, 25%)") should be rgb(63, 87, 39). Was rgb(64, 88, 40).
FAIL colorTest.parseColor("hsl(120, 37.5%, 25%)") should be rgb(39, 87, 39). Was rgb(40, 88, 40).
FAIL colorTest.parseColor("hsl(150, 37.5%, 25%)") should be rgb(39, 87, 63). Was rgb(40, 88, 64).
FAIL colorTest.parseColor("hsl(180, 37.5%, 25%)") should be rgb(39, 87, 87). Was rgb(40, 88, 88).
FAIL colorTest.parseColor("hsl(210, 37.5%, 25%)") should be rgb(39, 63, 87). Was rgb(40, 64, 88).
FAIL colorTest.parseColor("hsl(240, 37.5%, 25%)") should be rgb(39, 39, 87). Was rgb(40, 40, 88).
FAIL colorTest.parseColor("hsl(270, 37.5%, 25%)") should be rgb(63, 39, 87). Was rgb(64, 40, 88).
FAIL colorTest.parseColor("hsl(300, 37.5%, 25%)") should be rgb(87, 39, 87). Was rgb(88, 40, 88).
FAIL colorTest.parseColor("hsl(330, 37.5%, 25%)") should be rgb(87, 39, 63). Was rgb(88, 40, 64).
FAIL colorTest.parseColor("hsl(0, 50%, 25%)") should be rgb(95, 31, 31). Was rgb(96, 32, 32).
FAIL colorTest.parseColor("hsl(30, 50%, 25%)") should be rgb(95, 63, 31). Was rgb(96, 64, 32).
FAIL colorTest.parseColor("hsl(60, 50%, 25%)") should be rgb(95, 95, 31). Was rgb(96, 96, 32).
FAIL colorTest.parseColor("hsl(90, 50%, 25%)") should be rgb(63, 95, 31). Was rgb(64, 96, 32).
FAIL colorTest.parseColor("hsl(120, 50%, 25%)") should be rgb(31, 95, 31). Was rgb(32, 96, 32).
FAIL colorTest.parseColor("hsl(150, 50%, 25%)") should be rgb(31, 95, 63). Was rgb(32, 96, 64).
FAIL colorTest.parseColor("hsl(180, 50%, 25%)") should be rgb(31, 95, 95). Was rgb(32, 96, 96).
FAIL colorTest.parseColor("hsl(210, 50%, 25%)") should be rgb(31, 63, 95). Was rgb(32, 64, 96).
FAIL colorTest.parseColor("hsl(240, 50%, 25%)") should be rgb(31, 31, 95). Was rgb(32, 32, 96).
FAIL colorTest.parseColor("hsl(270, 50%, 25%)") should be rgb(63, 31, 95). Was rgb(64, 32, 96).
FAIL colorTest.parseColor("hsl(300, 50%, 25%)") should be rgb(95, 31, 95). Was rgb(96, 32, 96).
FAIL colorTest.parseColor("hsl(330, 50%, 25%)") should be rgb(95, 31, 63). Was rgb(96, 32, 64).
FAIL colorTest.parseColor("hsl(0, 62.5%, 25%)") should be rgb(103, 23, 23). Was rgb(104, 24, 24).
FAIL colorTest.parseColor("hsl(30, 62.5%, 25%)") should be rgb(103, 63, 23). Was rgb(104, 64, 24).
FAIL colorTest.parseColor("hsl(60, 62.5%, 25%)") should be rgb(103, 103, 23). Was rgb(104, 104, 24).
FAIL colorTest.parseColor("hsl(90, 62.5%, 25%)") should be rgb(63, 103, 23). Was rgb(64, 104, 24).
FAIL colorTest.parseColor("hsl(120, 62.5%, 25%)") should be rgb(23, 103, 23). Was rgb(24, 104, 24).
FAIL colorTest.parseColor("hsl(150, 62.5%, 25%)") should be rgb(23, 103, 63). Was rgb(24, 104, 64).
FAIL colorTest.parseColor("hsl(180, 62.5%, 25%)") should be rgb(23, 103, 103). Was rgb(24, 104, 104).
FAIL colorTest.parseColor("hsl(210, 62.5%, 25%)") should be rgb(23, 63, 103). Was rgb(24, 64, 104).
FAIL colorTest.parseColor("hsl(240, 62.5%, 25%)") should be rgb(23, 23, 103). Was rgb(24, 24, 104).
FAIL colorTest.parseColor("hsl(270, 62.5%, 25%)") should be rgb(63, 23, 103). Was rgb(64, 24, 104).
FAIL colorTest.parseColor("hsl(300, 62.5%, 25%)") should be rgb(103, 23, 103). Was rgb(104, 24, 104).
FAIL colorTest.parseColor("hsl(330, 62.5%, 25%)") should be rgb(103, 23, 63). Was rgb(104, 24, 64).
FAIL colorTest.parseColor("hsl(0, 75%, 25%)") should be rgb(111, 15, 15). Was rgb(112, 16, 16).
FAIL colorTest.parseColor("hsl(30, 75%, 25%)") should be rgb(111, 63, 15). Was rgb(112, 64, 16).
FAIL colorTest.parseColor("hsl(60, 75%, 25%)") should be rgb(111, 111, 15). Was rgb(112, 112, 16).
FAIL colorTest.parseColor("hsl(90, 75%, 25%)") should be rgb(63, 111, 15). Was rgb(64, 112, 16).
FAIL colorTest.parseColor("hsl(120, 75%, 25%)") should be rgb(15, 111, 15). Was rgb(16, 112, 16).
FAIL colorTest.parseColor("hsl(150, 75%, 25%)") should be rgb(15, 111, 63). Was rgb(16, 112, 64).
FAIL colorTest.parseColor("hsl(180, 75%, 25%)") should be rgb(15, 111, 111). Was rgb(16, 112, 112).
FAIL colorTest.parseColor("hsl(210, 75%, 25%)") should be rgb(15, 63, 111). Was rgb(16, 64, 112).
FAIL colorTest.parseColor("hsl(240, 75%, 25%)") should be rgb(15, 15, 111). Was rgb(16, 16, 112).
FAIL colorTest.parseColor("hsl(270, 75%, 25%)") should be rgb(63, 15, 111). Was rgb(64, 16, 112).
FAIL colorTest.parseColor("hsl(300, 75%, 25%)") should be rgb(111, 15, 111). Was rgb(112, 16, 112).
FAIL colorTest.parseColor("hsl(330, 75%, 25%)") should be rgb(111, 15, 63). Was rgb(112, 16, 64).
FAIL colorTest.parseColor("hsl(0, 87.5%, 25%)") should be rgb(119, 7, 7). Was rgb(120, 8, 8).
FAIL colorTest.parseColor("hsl(30, 87.5%, 25%)") should be rgb(119, 63, 7). Was rgb(120, 64, 8).
FAIL colorTest.parseColor("hsl(60, 87.5%, 25%)") should be rgb(119, 119, 7). Was rgb(120, 120, 8).
FAIL colorTest.parseColor("hsl(90, 87.5%, 25%)") should be rgb(63, 119, 7). Was rgb(64, 120, 8).
FAIL colorTest.parseColor("hsl(120, 87.5%, 25%)") should be rgb(7, 119, 7). Was rgb(8, 120, 8).
FAIL colorTest.parseColor("hsl(150, 87.5%, 25%)") should be rgb(7, 119, 63). Was rgb(8, 120, 64).
FAIL colorTest.parseColor("hsl(180, 87.5%, 25%)") should be rgb(7, 119, 119). Was rgb(8, 120, 120).
FAIL colorTest.parseColor("hsl(210, 87.5%, 25%)") should be rgb(7, 63, 119). Was rgb(8, 64, 120).
FAIL colorTest.parseColor("hsl(240, 87.5%, 25%)") should be rgb(7, 7, 119). Was rgb(8, 8, 120).
FAIL colorTest.parseColor("hsl(270, 87.5%, 25%)") should be rgb(63, 7, 119). Was rgb(64, 8, 120).
FAIL colorTest.parseColor("hsl(300, 87.5%, 25%)") should be rgb(119, 7, 119). Was rgb(120, 8, 120).
FAIL colorTest.parseColor("hsl(330, 87.5%, 25%)") should be rgb(119, 7, 63). Was rgb(120, 8, 64).
FAIL colorTest.parseColor("hsl(0, 100%, 25%)") should be rgb(127, 0, 0). Was rgb(128, 0, 0).
FAIL colorTest.parseColor("hsl(30, 100%, 25%)") should be rgb(127, 63, 0). Was rgb(128, 64, 0).
FAIL colorTest.parseColor("hsl(60, 100%, 25%)") should be rgb(127, 127, 0). Was rgb(128, 128, 0).
FAIL colorTest.parseColor("hsl(90, 100%, 25%)") should be rgb(63, 127, 0). Was rgb(64, 128, 0).
FAIL colorTest.parseColor("hsl(120, 100%, 25%)") should be rgb(0, 127, 0). Was rgb(0, 128, 0).
FAIL colorTest.parseColor("hsl(150, 100%, 25%)") should be rgb(0, 127, 63). Was rgb(0, 128, 64).
FAIL colorTest.parseColor("hsl(180, 100%, 25%)") should be rgb(0, 127, 127). Was rgb(0, 128, 128).
FAIL colorTest.parseColor("hsl(210, 100%, 25%)") should be rgb(0, 63, 127). Was rgb(0, 64, 128).
FAIL colorTest.parseColor("hsl(240, 100%, 25%)") should be rgb(0, 0, 127). Was rgb(0, 0, 128).
FAIL colorTest.parseColor("hsl(270, 100%, 25%)") should be rgb(63, 0, 127). Was rgb(64, 0, 128).
FAIL colorTest.parseColor("hsl(300, 100%, 25%)") should be rgb(127, 0, 127). Was rgb(128, 0, 128).
FAIL colorTest.parseColor("hsl(330, 100%, 25%)") should be rgb(127, 0, 63). Was rgb(128, 0, 64).
FAIL colorTest.parseColor("hsl(0, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(30, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(60, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(90, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(120, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(150, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(180, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(210, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(240, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(270, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(300, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(330, 0%, 37.5%)") should be rgb(95, 95, 95). Was rgb(96, 96, 96).
FAIL colorTest.parseColor("hsl(0, 12.5%, 37.5%)") should be rgb(107, 83, 83). Was rgb(108, 84, 84).
FAIL colorTest.parseColor("hsl(30, 12.5%, 37.5%)") should be rgb(107, 95, 83). Was rgb(108, 96, 84).
FAIL colorTest.parseColor("hsl(60, 12.5%, 37.5%)") should be rgb(107, 107, 83). Was rgb(108, 108, 84).
FAIL colorTest.parseColor("hsl(90, 12.5%, 37.5%)") should be rgb(95, 107, 83). Was rgb(96, 108, 84).
FAIL colorTest.parseColor("hsl(120, 12.5%, 37.5%)") should be rgb(83, 107, 83). Was rgb(84, 108, 84).
FAIL colorTest.parseColor("hsl(150, 12.5%, 37.5%)") should be rgb(83, 107, 95). Was rgb(84, 108, 96).
FAIL colorTest.parseColor("hsl(180, 12.5%, 37.5%)") should be rgb(83, 107, 107). Was rgb(84, 108, 108).
FAIL colorTest.parseColor("hsl(210, 12.5%, 37.5%)") should be rgb(83, 95, 107). Was rgb(84, 96, 108).
FAIL colorTest.parseColor("hsl(240, 12.5%, 37.5%)") should be rgb(83, 83, 107). Was rgb(84, 84, 108).
FAIL colorTest.parseColor("hsl(270, 12.5%, 37.5%)") should be rgb(95, 83, 107). Was rgb(96, 84, 108).
FAIL colorTest.parseColor("hsl(300, 12.5%, 37.5%)") should be rgb(107, 83, 107). Was rgb(108, 84, 108).
FAIL colorTest.parseColor("hsl(330, 12.5%, 37.5%)") should be rgb(107, 83, 95). Was rgb(108, 84, 96).
FAIL colorTest.parseColor("hsl(0, 25%, 37.5%)") should be rgb(119, 71, 71). Was rgb(120, 72, 72).
FAIL colorTest.parseColor("hsl(30, 25%, 37.5%)") should be rgb(119, 95, 71). Was rgb(120, 96, 72).
FAIL colorTest.parseColor("hsl(60, 25%, 37.5%)") should be rgb(119, 119, 71). Was rgb(120, 120, 72).
FAIL colorTest.parseColor("hsl(90, 25%, 37.5%)") should be rgb(95, 119, 71). Was rgb(96, 120, 72).
FAIL colorTest.parseColor("hsl(120, 25%, 37.5%)") should be rgb(71, 119, 71). Was rgb(72, 120, 72).
FAIL colorTest.parseColor("hsl(150, 25%, 37.5%)") should be rgb(71, 119, 95). Was rgb(72, 120, 96).
FAIL colorTest.parseColor("hsl(180, 25%, 37.5%)") should be rgb(71, 119, 119). Was rgb(72, 120, 120).
FAIL colorTest.parseColor("hsl(210, 25%, 37.5%)") should be rgb(71, 95, 119). Was rgb(72, 96, 120).
FAIL colorTest.parseColor("hsl(240, 25%, 37.5%)") should be rgb(71, 71, 119). Was rgb(72, 72, 120).
FAIL colorTest.parseColor("hsl(270, 25%, 37.5%)") should be rgb(95, 71, 119). Was rgb(96, 72, 120).
FAIL colorTest.parseColor("hsl(300, 25%, 37.5%)") should be rgb(119, 71, 119). Was rgb(120, 72, 120).
FAIL colorTest.parseColor("hsl(330, 25%, 37.5%)") should be rgb(119, 71, 95). Was rgb(120, 72, 96).
FAIL colorTest.parseColor("hsl(0, 37.5%, 37.5%)") should be rgb(131, 59, 59). Was rgb(131, 60, 60).
FAIL colorTest.parseColor("hsl(30, 37.5%, 37.5%)") should be rgb(131, 95, 59). Was rgb(131, 96, 60).
FAIL colorTest.parseColor("hsl(60, 37.5%, 37.5%)") should be rgb(131, 131, 59). Was rgb(131, 131, 60).
FAIL colorTest.parseColor("hsl(90, 37.5%, 37.5%)") should be rgb(95, 131, 59). Was rgb(96, 131, 60).
FAIL colorTest.parseColor("hsl(120, 37.5%, 37.5%)") should be rgb(59, 131, 59). Was rgb(60, 131, 60).
FAIL colorTest.parseColor("hsl(150, 37.5%, 37.5%)") should be rgb(59, 131, 95). Was rgb(60, 131, 96).
FAIL colorTest.parseColor("hsl(180, 37.5%, 37.5%)") should be rgb(59, 131, 131). Was rgb(60, 131, 131).
FAIL colorTest.parseColor("hsl(210, 37.5%, 37.5%)") should be rgb(59, 95, 131). Was rgb(60, 96, 131).
FAIL colorTest.parseColor("hsl(240, 37.5%, 37.5%)") should be rgb(59, 59, 131). Was rgb(60, 60, 131).
FAIL colorTest.parseColor("hsl(270, 37.5%, 37.5%)") should be rgb(95, 59, 131). Was rgb(96, 60, 131).
FAIL colorTest.parseColor("hsl(300, 37.5%, 37.5%)") should be rgb(131, 59, 131). Was rgb(131, 60, 131).
FAIL colorTest.parseColor("hsl(330, 37.5%, 37.5%)") should be rgb(131, 59, 95). Was rgb(131, 60, 96).
FAIL colorTest.parseColor("hsl(0, 50%, 37.5%)") should be rgb(143, 47, 47). Was rgb(143, 48, 48).
FAIL colorTest.parseColor("hsl(30, 50%, 37.5%)") should be rgb(143, 95, 47). Was rgb(143, 96, 48).
FAIL colorTest.parseColor("hsl(60, 50%, 37.5%)") should be rgb(143, 143, 47). Was rgb(143, 143, 48).
FAIL colorTest.parseColor("hsl(90, 50%, 37.5%)") should be rgb(95, 143, 47). Was rgb(96, 143, 48).
FAIL colorTest.parseColor("hsl(120, 50%, 37.5%)") should be rgb(47, 143, 47). Was rgb(48, 143, 48).
FAIL colorTest.parseColor("hsl(150, 50%, 37.5%)") should be rgb(47, 143, 95). Was rgb(48, 143, 96).
FAIL colorTest.parseColor("hsl(180, 50%, 37.5%)") should be rgb(47, 143, 143). Was rgb(48, 143, 143).
FAIL colorTest.parseColor("hsl(210, 50%, 37.5%)") should be rgb(47, 95, 143). Was rgb(48, 96, 143).
FAIL colorTest.parseColor("hsl(240, 50%, 37.5%)") should be rgb(47, 47, 143). Was rgb(48, 48, 143).
FAIL colorTest.parseColor("hsl(270, 50%, 37.5%)") should be rgb(95, 47, 143). Was rgb(96, 48, 143).
FAIL colorTest.parseColor("hsl(300, 50%, 37.5%)") should be rgb(143, 47, 143). Was rgb(143, 48, 143).
FAIL colorTest.parseColor("hsl(330, 50%, 37.5%)") should be rgb(143, 47, 95). Was rgb(143, 48, 96).
FAIL colorTest.parseColor("hsl(0, 62.5%, 37.5%)") should be rgb(155, 35, 35). Was rgb(155, 36, 36).
FAIL colorTest.parseColor("hsl(30, 62.5%, 37.5%)") should be rgb(155, 95, 35). Was rgb(155, 96, 36).
FAIL colorTest.parseColor("hsl(60, 62.5%, 37.5%)") should be rgb(155, 155, 35). Was rgb(155, 155, 36).
FAIL colorTest.parseColor("hsl(90, 62.5%, 37.5%)") should be rgb(95, 155, 35). Was rgb(96, 155, 36).
FAIL colorTest.parseColor("hsl(120, 62.5%, 37.5%)") should be rgb(35, 155, 35). Was rgb(36, 155, 36).
FAIL colorTest.parseColor("hsl(150, 62.5%, 37.5%)") should be rgb(35, 155, 95). Was rgb(36, 155, 96).
FAIL colorTest.parseColor("hsl(180, 62.5%, 37.5%)") should be rgb(35, 155, 155). Was rgb(36, 155, 155).
FAIL colorTest.parseColor("hsl(210, 62.5%, 37.5%)") should be rgb(35, 95, 155). Was rgb(36, 96, 155).
FAIL colorTest.parseColor("hsl(240, 62.5%, 37.5%)") should be rgb(35, 35, 155). Was rgb(36, 36, 155).
FAIL colorTest.parseColor("hsl(270, 62.5%, 37.5%)") should be rgb(95, 35, 155). Was rgb(96, 36, 155).
FAIL colorTest.parseColor("hsl(300, 62.5%, 37.5%)") should be rgb(155, 35, 155). Was rgb(155, 36, 155).
FAIL colorTest.parseColor("hsl(330, 62.5%, 37.5%)") should be rgb(155, 35, 95). Was rgb(155, 36, 96).
FAIL colorTest.parseColor("hsl(0, 75%, 37.5%)") should be rgb(167, 23, 23). Was rgb(167, 24, 24).
FAIL colorTest.parseColor("hsl(30, 75%, 37.5%)") should be rgb(167, 95, 23). Was rgb(167, 96, 24).
FAIL colorTest.parseColor("hsl(60, 75%, 37.5%)") should be rgb(167, 167, 23). Was rgb(167, 167, 24).
FAIL colorTest.parseColor("hsl(90, 75%, 37.5%)") should be rgb(95, 167, 23). Was rgb(96, 167, 24).
FAIL colorTest.parseColor("hsl(120, 75%, 37.5%)") should be rgb(23, 167, 23). Was rgb(24, 167, 24).
FAIL colorTest.parseColor("hsl(150, 75%, 37.5%)") should be rgb(23, 167, 95). Was rgb(24, 167, 96).
FAIL colorTest.parseColor("hsl(180, 75%, 37.5%)") should be rgb(23, 167, 167). Was rgb(24, 167, 167).
FAIL colorTest.parseColor("hsl(210, 75%, 37.5%)") should be rgb(23, 95, 167). Was rgb(24, 96, 167).
FAIL colorTest.parseColor("hsl(240, 75%, 37.5%)") should be rgb(23, 23, 167). Was rgb(24, 24, 167).
FAIL colorTest.parseColor("hsl(270, 75%, 37.5%)") should be rgb(95, 23, 167). Was rgb(96, 24, 167).
FAIL colorTest.parseColor("hsl(300, 75%, 37.5%)") should be rgb(167, 23, 167). Was rgb(167, 24, 167).
FAIL colorTest.parseColor("hsl(330, 75%, 37.5%)") should be rgb(167, 23, 95). Was rgb(167, 24, 96).
FAIL colorTest.parseColor("hsl(0, 87.5%, 37.5%)") should be rgb(179, 11, 11). Was rgb(179, 12, 12).
FAIL colorTest.parseColor("hsl(30, 87.5%, 37.5%)") should be rgb(179, 95, 11). Was rgb(179, 96, 12).
FAIL colorTest.parseColor("hsl(60, 87.5%, 37.5%)") should be rgb(179, 179, 11). Was rgb(179, 179, 12).
FAIL colorTest.parseColor("hsl(90, 87.5%, 37.5%)") should be rgb(95, 179, 11). Was rgb(96, 179, 12).
FAIL colorTest.parseColor("hsl(120, 87.5%, 37.5%)") should be rgb(11, 179, 11). Was rgb(12, 179, 12).
FAIL colorTest.parseColor("hsl(150, 87.5%, 37.5%)") should be rgb(11, 179, 95). Was rgb(12, 179, 96).
FAIL colorTest.parseColor("hsl(180, 87.5%, 37.5%)") should be rgb(11, 179, 179). Was rgb(12, 179, 179).
FAIL colorTest.parseColor("hsl(210, 87.5%, 37.5%)") should be rgb(11, 95, 179). Was rgb(12, 96, 179).
FAIL colorTest.parseColor("hsl(240, 87.5%, 37.5%)") should be rgb(11, 11, 179). Was rgb(12, 12, 179).
FAIL colorTest.parseColor("hsl(270, 87.5%, 37.5%)") should be rgb(95, 11, 179). Was rgb(96, 12, 179).
FAIL colorTest.parseColor("hsl(300, 87.5%, 37.5%)") should be rgb(179, 11, 179). Was rgb(179, 12, 179).
FAIL colorTest.parseColor("hsl(330, 87.5%, 37.5%)") should be rgb(179, 11, 95). Was rgb(179, 12, 96).
PASS colorTest.parseColor("hsl(0, 100%, 37.5%)") is "rgb(191, 0, 0)"
FAIL colorTest.parseColor("hsl(30, 100%, 37.5%)") should be rgb(191, 95, 0). Was rgb(191, 96, 0).
PASS colorTest.parseColor("hsl(60, 100%, 37.5%)") is "rgb(191, 191, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 37.5%)") should be rgb(95, 191, 0). Was rgb(96, 191, 0).
PASS colorTest.parseColor("hsl(120, 100%, 37.5%)") is "rgb(0, 191, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 37.5%)") should be rgb(0, 191, 95). Was rgb(0, 191, 96).
PASS colorTest.parseColor("hsl(180, 100%, 37.5%)") is "rgb(0, 191, 191)"
FAIL colorTest.parseColor("hsl(210, 100%, 37.5%)") should be rgb(0, 95, 191). Was rgb(0, 96, 191).
PASS colorTest.parseColor("hsl(240, 100%, 37.5%)") is "rgb(0, 0, 191)"
FAIL colorTest.parseColor("hsl(270, 100%, 37.5%)") should be rgb(95, 0, 191). Was rgb(96, 0, 191).
PASS colorTest.parseColor("hsl(300, 100%, 37.5%)") is "rgb(191, 0, 191)"
FAIL colorTest.parseColor("hsl(330, 100%, 37.5%)") should be rgb(191, 0, 95). Was rgb(191, 0, 96).
FAIL colorTest.parseColor("hsl(0, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(30, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(60, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(90, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(120, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(150, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(180, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(210, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(240, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(270, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(300, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(330, 0%, 50%)") should be rgb(127, 127, 127). Was rgb(128, 128, 128).
FAIL colorTest.parseColor("hsl(0, 12.5%, 50%)") should be rgb(143, 111, 111). Was rgb(143, 112, 112).
FAIL colorTest.parseColor("hsl(30, 12.5%, 50%)") should be rgb(143, 127, 111). Was rgb(143, 128, 112).
FAIL colorTest.parseColor("hsl(60, 12.5%, 50%)") should be rgb(143, 143, 111). Was rgb(143, 143, 112).
FAIL colorTest.parseColor("hsl(90, 12.5%, 50%)") should be rgb(127, 143, 111). Was rgb(128, 143, 112).
FAIL colorTest.parseColor("hsl(120, 12.5%, 50%)") should be rgb(111, 143, 111). Was rgb(112, 143, 112).
FAIL colorTest.parseColor("hsl(150, 12.5%, 50%)") should be rgb(111, 143, 127). Was rgb(112, 143, 128).
FAIL colorTest.parseColor("hsl(180, 12.5%, 50%)") should be rgb(111, 143, 143). Was rgb(112, 143, 143).
FAIL colorTest.parseColor("hsl(210, 12.5%, 50%)") should be rgb(111, 127, 143). Was rgb(112, 128, 143).
FAIL colorTest.parseColor("hsl(240, 12.5%, 50%)") should be rgb(111, 111, 143). Was rgb(112, 112, 143).
FAIL colorTest.parseColor("hsl(270, 12.5%, 50%)") should be rgb(127, 111, 143). Was rgb(128, 112, 143).
FAIL colorTest.parseColor("hsl(300, 12.5%, 50%)") should be rgb(143, 111, 143). Was rgb(143, 112, 143).
FAIL colorTest.parseColor("hsl(330, 12.5%, 50%)") should be rgb(143, 111, 127). Was rgb(143, 112, 128).
FAIL colorTest.parseColor("hsl(0, 25%, 50%)") should be rgb(159, 95, 95). Was rgb(159, 96, 96).
FAIL colorTest.parseColor("hsl(30, 25%, 50%)") should be rgb(159, 127, 95). Was rgb(159, 128, 96).
FAIL colorTest.parseColor("hsl(60, 25%, 50%)") should be rgb(159, 159, 95). Was rgb(159, 159, 96).
FAIL colorTest.parseColor("hsl(90, 25%, 50%)") should be rgb(127, 159, 95). Was rgb(128, 159, 96).
FAIL colorTest.parseColor("hsl(120, 25%, 50%)") should be rgb(95, 159, 95). Was rgb(96, 159, 96).
FAIL colorTest.parseColor("hsl(150, 25%, 50%)") should be rgb(95, 159, 127). Was rgb(96, 159, 128).
FAIL colorTest.parseColor("hsl(180, 25%, 50%)") should be rgb(95, 159, 159). Was rgb(96, 159, 159).
FAIL colorTest.parseColor("hsl(210, 25%, 50%)") should be rgb(95, 127, 159). Was rgb(96, 128, 159).
FAIL colorTest.parseColor("hsl(240, 25%, 50%)") should be rgb(95, 95, 159). Was rgb(96, 96, 159).
FAIL colorTest.parseColor("hsl(270, 25%, 50%)") should be rgb(127, 95, 159). Was rgb(128, 96, 159).
FAIL colorTest.parseColor("hsl(300, 25%, 50%)") should be rgb(159, 95, 159). Was rgb(159, 96, 159).
FAIL colorTest.parseColor("hsl(330, 25%, 50%)") should be rgb(159, 95, 127). Was rgb(159, 96, 128).
FAIL colorTest.parseColor("hsl(0, 37.5%, 50%)") should be rgb(175, 79, 79). Was rgb(175, 80, 80).
FAIL colorTest.parseColor("hsl(30, 37.5%, 50%)") should be rgb(175, 127, 79). Was rgb(175, 128, 80).
FAIL colorTest.parseColor("hsl(60, 37.5%, 50%)") should be rgb(175, 175, 79). Was rgb(175, 175, 80).
FAIL colorTest.parseColor("hsl(90, 37.5%, 50%)") should be rgb(127, 175, 79). Was rgb(128, 175, 80).
FAIL colorTest.parseColor("hsl(120, 37.5%, 50%)") should be rgb(79, 175, 79). Was rgb(80, 175, 80).
FAIL colorTest.parseColor("hsl(150, 37.5%, 50%)") should be rgb(79, 175, 127). Was rgb(80, 175, 128).
FAIL colorTest.parseColor("hsl(180, 37.5%, 50%)") should be rgb(79, 175, 175). Was rgb(80, 175, 175).
FAIL colorTest.parseColor("hsl(210, 37.5%, 50%)") should be rgb(79, 127, 175). Was rgb(80, 128, 175).
FAIL colorTest.parseColor("hsl(240, 37.5%, 50%)") should be rgb(79, 79, 175). Was rgb(80, 80, 175).
FAIL colorTest.parseColor("hsl(270, 37.5%, 50%)") should be rgb(127, 79, 175). Was rgb(128, 80, 175).
FAIL colorTest.parseColor("hsl(300, 37.5%, 50%)") should be rgb(175, 79, 175). Was rgb(175, 80, 175).
FAIL colorTest.parseColor("hsl(330, 37.5%, 50%)") should be rgb(175, 79, 127). Was rgb(175, 80, 128).
FAIL colorTest.parseColor("hsl(0, 50%, 50%)") should be rgb(191, 63, 63). Was rgb(191, 64, 64).
FAIL colorTest.parseColor("hsl(30, 50%, 50%)") should be rgb(191, 127, 63). Was rgb(191, 128, 64).
FAIL colorTest.parseColor("hsl(60, 50%, 50%)") should be rgb(191, 191, 63). Was rgb(191, 191, 64).
FAIL colorTest.parseColor("hsl(90, 50%, 50%)") should be rgb(127, 191, 63). Was rgb(128, 191, 64).
FAIL colorTest.parseColor("hsl(120, 50%, 50%)") should be rgb(63, 191, 63). Was rgb(64, 191, 64).
FAIL colorTest.parseColor("hsl(150, 50%, 50%)") should be rgb(63, 191, 127). Was rgb(64, 191, 128).
FAIL colorTest.parseColor("hsl(180, 50%, 50%)") should be rgb(63, 191, 191). Was rgb(64, 191, 191).
FAIL colorTest.parseColor("hsl(210, 50%, 50%)") should be rgb(63, 127, 191). Was rgb(64, 128, 191).
FAIL colorTest.parseColor("hsl(240, 50%, 50%)") should be rgb(63, 63, 191). Was rgb(64, 64, 191).
FAIL colorTest.parseColor("hsl(270, 50%, 50%)") should be rgb(127, 63, 191). Was rgb(128, 64, 191).
FAIL colorTest.parseColor("hsl(300, 50%, 50%)") should be rgb(191, 63, 191). Was rgb(191, 64, 191).
FAIL colorTest.parseColor("hsl(330, 50%, 50%)") should be rgb(191, 63, 127). Was rgb(191, 64, 128).
FAIL colorTest.parseColor("hsl(0, 62.5%, 50%)") should be rgb(207, 47, 47). Was rgb(207, 48, 48).
FAIL colorTest.parseColor("hsl(30, 62.5%, 50%)") should be rgb(207, 127, 47). Was rgb(207, 128, 48).
FAIL colorTest.parseColor("hsl(60, 62.5%, 50%)") should be rgb(207, 207, 47). Was rgb(207, 207, 48).
FAIL colorTest.parseColor("hsl(90, 62.5%, 50%)") should be rgb(127, 207, 47). Was rgb(128, 207, 48).
FAIL colorTest.parseColor("hsl(120, 62.5%, 50%)") should be rgb(47, 207, 47). Was rgb(48, 207, 48).
FAIL colorTest.parseColor("hsl(150, 62.5%, 50%)") should be rgb(47, 207, 127). Was rgb(48, 207, 128).
FAIL colorTest.parseColor("hsl(180, 62.5%, 50%)") should be rgb(47, 207, 207). Was rgb(48, 207, 207).
FAIL colorTest.parseColor("hsl(210, 62.5%, 50%)") should be rgb(47, 127, 207). Was rgb(48, 128, 207).
FAIL colorTest.parseColor("hsl(240, 62.5%, 50%)") should be rgb(47, 47, 207). Was rgb(48, 48, 207).
FAIL colorTest.parseColor("hsl(270, 62.5%, 50%)") should be rgb(127, 47, 207). Was rgb(128, 48, 207).
FAIL colorTest.parseColor("hsl(300, 62.5%, 50%)") should be rgb(207, 47, 207). Was rgb(207, 48, 207).
FAIL colorTest.parseColor("hsl(330, 62.5%, 50%)") should be rgb(207, 47, 127). Was rgb(207, 48, 128).
FAIL colorTest.parseColor("hsl(0, 75%, 50%)") should be rgb(223, 31, 31). Was rgb(223, 32, 32).
FAIL colorTest.parseColor("hsl(30, 75%, 50%)") should be rgb(223, 127, 31). Was rgb(223, 128, 32).
FAIL colorTest.parseColor("hsl(60, 75%, 50%)") should be rgb(223, 223, 31). Was rgb(223, 223, 32).
FAIL colorTest.parseColor("hsl(90, 75%, 50%)") should be rgb(127, 223, 31). Was rgb(128, 223, 32).
FAIL colorTest.parseColor("hsl(120, 75%, 50%)") should be rgb(31, 223, 31). Was rgb(32, 223, 32).
FAIL colorTest.parseColor("hsl(150, 75%, 50%)") should be rgb(31, 223, 127). Was rgb(32, 223, 128).
FAIL colorTest.parseColor("hsl(180, 75%, 50%)") should be rgb(31, 223, 223). Was rgb(32, 223, 223).
FAIL colorTest.parseColor("hsl(210, 75%, 50%)") should be rgb(31, 127, 223). Was rgb(32, 128, 223).
FAIL colorTest.parseColor("hsl(240, 75%, 50%)") should be rgb(31, 31, 223). Was rgb(32, 32, 223).
FAIL colorTest.parseColor("hsl(270, 75%, 50%)") should be rgb(127, 31, 223). Was rgb(128, 32, 223).
FAIL colorTest.parseColor("hsl(300, 75%, 50%)") should be rgb(223, 31, 223). Was rgb(223, 32, 223).
FAIL colorTest.parseColor("hsl(330, 75%, 50%)") should be rgb(223, 31, 127). Was rgb(223, 32, 128).
FAIL colorTest.parseColor("hsl(0, 87.5%, 50%)") should be rgb(239, 15, 15). Was rgb(239, 16, 16).
FAIL colorTest.parseColor("hsl(30, 87.5%, 50%)") should be rgb(239, 127, 15). Was rgb(239, 128, 16).
FAIL colorTest.parseColor("hsl(60, 87.5%, 50%)") should be rgb(239, 239, 15). Was rgb(239, 239, 16).
FAIL colorTest.parseColor("hsl(90, 87.5%, 50%)") should be rgb(127, 239, 15). Was rgb(128, 239, 16).
FAIL colorTest.parseColor("hsl(120, 87.5%, 50%)") should be rgb(15, 239, 15). Was rgb(16, 239, 16).
FAIL colorTest.parseColor("hsl(150, 87.5%, 50%)") should be rgb(15, 239, 127). Was rgb(16, 239, 128).
FAIL colorTest.parseColor("hsl(180, 87.5%, 50%)") should be rgb(15, 239, 239). Was rgb(16, 239, 239).
FAIL colorTest.parseColor("hsl(210, 87.5%, 50%)") should be rgb(15, 127, 239). Was rgb(16, 128, 239).
FAIL colorTest.parseColor("hsl(240, 87.5%, 50%)") should be rgb(15, 15, 239). Was rgb(16, 16, 239).
FAIL colorTest.parseColor("hsl(270, 87.5%, 50%)") should be rgb(127, 15, 239). Was rgb(128, 16, 239).
FAIL colorTest.parseColor("hsl(300, 87.5%, 50%)") should be rgb(239, 15, 239). Was rgb(239, 16, 239).
FAIL colorTest.parseColor("hsl(330, 87.5%, 50%)") should be rgb(239, 15, 127). Was rgb(239, 16, 128).
PASS colorTest.parseColor("hsl(0, 100%, 50%)") is "rgb(255, 0, 0)"
FAIL colorTest.parseColor("hsl(30, 100%, 50%)") should be rgb(255, 127, 0). Was rgb(255, 128, 0).
PASS colorTest.parseColor("hsl(60, 100%, 50%)") is "rgb(255, 255, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 50%)") should be rgb(127, 255, 0). Was rgb(128, 255, 0).
PASS colorTest.parseColor("hsl(120, 100%, 50%)") is "rgb(0, 255, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 50%)") should be rgb(0, 255, 127). Was rgb(0, 255, 128).
PASS colorTest.parseColor("hsl(180, 100%, 50%)") is "rgb(0, 255, 255)"
FAIL colorTest.parseColor("hsl(210, 100%, 50%)") should be rgb(0, 127, 255). Was rgb(0, 128, 255).
PASS colorTest.parseColor("hsl(240, 100%, 50%)") is "rgb(0, 0, 255)"
FAIL colorTest.parseColor("hsl(270, 100%, 50%)") should be rgb(127, 0, 255). Was rgb(128, 0, 255).
PASS colorTest.parseColor("hsl(300, 100%, 50%)") is "rgb(255, 0, 255)"
FAIL colorTest.parseColor("hsl(330, 100%, 50%)") should be rgb(255, 0, 127). Was rgb(255, 0, 128).
PASS colorTest.parseColor("hsl(0, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(30, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(60, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(90, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(120, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(150, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(180, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(210, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(240, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(270, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(300, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(330, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(0, 12.5%, 62.5%)") is "rgb(171, 147, 147)"
PASS colorTest.parseColor("hsl(30, 12.5%, 62.5%)") is "rgb(171, 159, 147)"
PASS colorTest.parseColor("hsl(60, 12.5%, 62.5%)") is "rgb(171, 171, 147)"
PASS colorTest.parseColor("hsl(90, 12.5%, 62.5%)") is "rgb(159, 171, 147)"
PASS colorTest.parseColor("hsl(120, 12.5%, 62.5%)") is "rgb(147, 171, 147)"
PASS colorTest.parseColor("hsl(150, 12.5%, 62.5%)") is "rgb(147, 171, 159)"
PASS colorTest.parseColor("hsl(180, 12.5%, 62.5%)") is "rgb(147, 171, 171)"
PASS colorTest.parseColor("hsl(210, 12.5%, 62.5%)") is "rgb(147, 159, 171)"
PASS colorTest.parseColor("hsl(240, 12.5%, 62.5%)") is "rgb(147, 147, 171)"
PASS colorTest.parseColor("hsl(270, 12.5%, 62.5%)") is "rgb(159, 147, 171)"
PASS colorTest.parseColor("hsl(300, 12.5%, 62.5%)") is "rgb(171, 147, 171)"
PASS colorTest.parseColor("hsl(330, 12.5%, 62.5%)") is "rgb(171, 147, 159)"
PASS colorTest.parseColor("hsl(0, 25%, 62.5%)") is "rgb(183, 135, 135)"
PASS colorTest.parseColor("hsl(30, 25%, 62.5%)") is "rgb(183, 159, 135)"
PASS colorTest.parseColor("hsl(60, 25%, 62.5%)") is "rgb(183, 183, 135)"
PASS colorTest.parseColor("hsl(90, 25%, 62.5%)") is "rgb(159, 183, 135)"
PASS colorTest.parseColor("hsl(120, 25%, 62.5%)") is "rgb(135, 183, 135)"
PASS colorTest.parseColor("hsl(150, 25%, 62.5%)") is "rgb(135, 183, 159)"
PASS colorTest.parseColor("hsl(180, 25%, 62.5%)") is "rgb(135, 183, 183)"
PASS colorTest.parseColor("hsl(210, 25%, 62.5%)") is "rgb(135, 159, 183)"
PASS colorTest.parseColor("hsl(240, 25%, 62.5%)") is "rgb(135, 135, 183)"
PASS colorTest.parseColor("hsl(270, 25%, 62.5%)") is "rgb(159, 135, 183)"
PASS colorTest.parseColor("hsl(300, 25%, 62.5%)") is "rgb(183, 135, 183)"
PASS colorTest.parseColor("hsl(330, 25%, 62.5%)") is "rgb(183, 135, 159)"
FAIL colorTest.parseColor("hsl(0, 37.5%, 62.5%)") should be rgb(195, 123, 123). Was rgb(195, 124, 124).
FAIL colorTest.parseColor("hsl(30, 37.5%, 62.5%)") should be rgb(195, 159, 123). Was rgb(195, 159, 124).
FAIL colorTest.parseColor("hsl(60, 37.5%, 62.5%)") should be rgb(195, 195, 123). Was rgb(195, 195, 124).
FAIL colorTest.parseColor("hsl(90, 37.5%, 62.5%)") should be rgb(159, 195, 123). Was rgb(159, 195, 124).
FAIL colorTest.parseColor("hsl(120, 37.5%, 62.5%)") should be rgb(123, 195, 123). Was rgb(124, 195, 124).
FAIL colorTest.parseColor("hsl(150, 37.5%, 62.5%)") should be rgb(123, 195, 159). Was rgb(124, 195, 159).
FAIL colorTest.parseColor("hsl(180, 37.5%, 62.5%)") should be rgb(123, 195, 195). Was rgb(124, 195, 195).
FAIL colorTest.parseColor("hsl(210, 37.5%, 62.5%)") should be rgb(123, 159, 195). Was rgb(124, 159, 195).
FAIL colorTest.parseColor("hsl(240, 37.5%, 62.5%)") should be rgb(123, 123, 195). Was rgb(124, 124, 195).
FAIL colorTest.parseColor("hsl(270, 37.5%, 62.5%)") should be rgb(159, 123, 195). Was rgb(159, 124, 195).
FAIL colorTest.parseColor("hsl(300, 37.5%, 62.5%)") should be rgb(195, 123, 195). Was rgb(195, 124, 195).
FAIL colorTest.parseColor("hsl(330, 37.5%, 62.5%)") should be rgb(195, 123, 159). Was rgb(195, 124, 159).
FAIL colorTest.parseColor("hsl(0, 50%, 62.5%)") should be rgb(207, 111, 111). Was rgb(207, 112, 112).
FAIL colorTest.parseColor("hsl(30, 50%, 62.5%)") should be rgb(207, 159, 111). Was rgb(207, 159, 112).
FAIL colorTest.parseColor("hsl(60, 50%, 62.5%)") should be rgb(207, 207, 111). Was rgb(207, 207, 112).
FAIL colorTest.parseColor("hsl(90, 50%, 62.5%)") should be rgb(159, 207, 111). Was rgb(159, 207, 112).
FAIL colorTest.parseColor("hsl(120, 50%, 62.5%)") should be rgb(111, 207, 111). Was rgb(112, 207, 112).
FAIL colorTest.parseColor("hsl(150, 50%, 62.5%)") should be rgb(111, 207, 159). Was rgb(112, 207, 159).
FAIL colorTest.parseColor("hsl(180, 50%, 62.5%)") should be rgb(111, 207, 207). Was rgb(112, 207, 207).
FAIL colorTest.parseColor("hsl(210, 50%, 62.5%)") should be rgb(111, 159, 207). Was rgb(112, 159, 207).
FAIL colorTest.parseColor("hsl(240, 50%, 62.5%)") should be rgb(111, 111, 207). Was rgb(112, 112, 207).
FAIL colorTest.parseColor("hsl(270, 50%, 62.5%)") should be rgb(159, 111, 207). Was rgb(159, 112, 207).
FAIL colorTest.parseColor("hsl(300, 50%, 62.5%)") should be rgb(207, 111, 207). Was rgb(207, 112, 207).
FAIL colorTest.parseColor("hsl(330, 50%, 62.5%)") should be rgb(207, 111, 159). Was rgb(207, 112, 159).
FAIL colorTest.parseColor("hsl(0, 62.5%, 62.5%)") should be rgb(219, 99, 99). Was rgb(219, 100, 100).
FAIL colorTest.parseColor("hsl(30, 62.5%, 62.5%)") should be rgb(219, 159, 99). Was rgb(219, 159, 100).
FAIL colorTest.parseColor("hsl(60, 62.5%, 62.5%)") should be rgb(219, 219, 99). Was rgb(219, 219, 100).
FAIL colorTest.parseColor("hsl(90, 62.5%, 62.5%)") should be rgb(159, 219, 99). Was rgb(159, 219, 100).
FAIL colorTest.parseColor("hsl(120, 62.5%, 62.5%)") should be rgb(99, 219, 99). Was rgb(100, 219, 100).
FAIL colorTest.parseColor("hsl(150, 62.5%, 62.5%)") should be rgb(99, 219, 159). Was rgb(100, 219, 159).
FAIL colorTest.parseColor("hsl(180, 62.5%, 62.5%)") should be rgb(99, 219, 219). Was rgb(100, 219, 219).
FAIL colorTest.parseColor("hsl(210, 62.5%, 62.5%)") should be rgb(99, 159, 219). Was rgb(100, 159, 219).
FAIL colorTest.parseColor("hsl(240, 62.5%, 62.5%)") should be rgb(99, 99, 219). Was rgb(100, 100, 219).
FAIL colorTest.parseColor("hsl(270, 62.5%, 62.5%)") should be rgb(159, 99, 219). Was rgb(159, 100, 219).
FAIL colorTest.parseColor("hsl(300, 62.5%, 62.5%)") should be rgb(219, 99, 219). Was rgb(219, 100, 219).
FAIL colorTest.parseColor("hsl(330, 62.5%, 62.5%)") should be rgb(219, 99, 159). Was rgb(219, 100, 159).
FAIL colorTest.parseColor("hsl(0, 75%, 62.5%)") should be rgb(231, 87, 87). Was rgb(231, 88, 88).
FAIL colorTest.parseColor("hsl(30, 75%, 62.5%)") should be rgb(231, 159, 87). Was rgb(231, 159, 88).
FAIL colorTest.parseColor("hsl(60, 75%, 62.5%)") should be rgb(231, 231, 87). Was rgb(231, 231, 88).
FAIL colorTest.parseColor("hsl(90, 75%, 62.5%)") should be rgb(159, 231, 87). Was rgb(159, 231, 88).
FAIL colorTest.parseColor("hsl(120, 75%, 62.5%)") should be rgb(87, 231, 87). Was rgb(88, 231, 88).
FAIL colorTest.parseColor("hsl(150, 75%, 62.5%)") should be rgb(87, 231, 159). Was rgb(88, 231, 159).
FAIL colorTest.parseColor("hsl(180, 75%, 62.5%)") should be rgb(87, 231, 231). Was rgb(88, 231, 231).
FAIL colorTest.parseColor("hsl(210, 75%, 62.5%)") should be rgb(87, 159, 231). Was rgb(88, 159, 231).
FAIL colorTest.parseColor("hsl(240, 75%, 62.5%)") should be rgb(87, 87, 231). Was rgb(88, 88, 231).
FAIL colorTest.parseColor("hsl(270, 75%, 62.5%)") should be rgb(159, 87, 231). Was rgb(159, 88, 231).
FAIL colorTest.parseColor("hsl(300, 75%, 62.5%)") should be rgb(231, 87, 231). Was rgb(231, 88, 231).
FAIL colorTest.parseColor("hsl(330, 75%, 62.5%)") should be rgb(231, 87, 159). Was rgb(231, 88, 159).
FAIL colorTest.parseColor("hsl(0, 87.5%, 62.5%)") should be rgb(243, 75, 75). Was rgb(243, 76, 76).
FAIL colorTest.parseColor("hsl(30, 87.5%, 62.5%)") should be rgb(243, 159, 75). Was rgb(243, 159, 76).
FAIL colorTest.parseColor("hsl(60, 87.5%, 62.5%)") should be rgb(243, 243, 75). Was rgb(243, 243, 76).
FAIL colorTest.parseColor("hsl(90, 87.5%, 62.5%)") should be rgb(159, 243, 75). Was rgb(159, 243, 76).
FAIL colorTest.parseColor("hsl(120, 87.5%, 62.5%)") should be rgb(75, 243, 75). Was rgb(76, 243, 76).
FAIL colorTest.parseColor("hsl(150, 87.5%, 62.5%)") should be rgb(75, 243, 159). Was rgb(76, 243, 159).
FAIL colorTest.parseColor("hsl(180, 87.5%, 62.5%)") should be rgb(75, 243, 243). Was rgb(76, 243, 243).
FAIL colorTest.parseColor("hsl(210, 87.5%, 62.5%)") should be rgb(75, 159, 243). Was rgb(76, 159, 243).
FAIL colorTest.parseColor("hsl(240, 87.5%, 62.5%)") should be rgb(75, 75, 243). Was rgb(76, 76, 243).
FAIL colorTest.parseColor("hsl(270, 87.5%, 62.5%)") should be rgb(159, 75, 243). Was rgb(159, 76, 243).
FAIL colorTest.parseColor("hsl(300, 87.5%, 62.5%)") should be rgb(243, 75, 243). Was rgb(243, 76, 243).
FAIL colorTest.parseColor("hsl(330, 87.5%, 62.5%)") should be rgb(243, 75, 159). Was rgb(243, 76, 159).
FAIL colorTest.parseColor("hsl(0, 100%, 62.5%)") should be rgb(255, 63, 63). Was rgb(255, 64, 64).
FAIL colorTest.parseColor("hsl(30, 100%, 62.5%)") should be rgb(255, 159, 63). Was rgb(255, 159, 64).
FAIL colorTest.parseColor("hsl(60, 100%, 62.5%)") should be rgb(255, 255, 63). Was rgb(255, 255, 64).
FAIL colorTest.parseColor("hsl(90, 100%, 62.5%)") should be rgb(159, 255, 63). Was rgb(159, 255, 64).
FAIL colorTest.parseColor("hsl(120, 100%, 62.5%)") should be rgb(63, 255, 63). Was rgb(64, 255, 64).
FAIL colorTest.parseColor("hsl(150, 100%, 62.5%)") should be rgb(63, 255, 159). Was rgb(64, 255, 159).
FAIL colorTest.parseColor("hsl(180, 100%, 62.5%)") should be rgb(63, 255, 255). Was rgb(64, 255, 255).
FAIL colorTest.parseColor("hsl(210, 100%, 62.5%)") should be rgb(63, 159, 255). Was rgb(64, 159, 255).
FAIL colorTest.parseColor("hsl(240, 100%, 62.5%)") should be rgb(63, 63, 255). Was rgb(64, 64, 255).
FAIL colorTest.parseColor("hsl(270, 100%, 62.5%)") should be rgb(159, 63, 255). Was rgb(159, 64, 255).
FAIL colorTest.parseColor("hsl(300, 100%, 62.5%)") should be rgb(255, 63, 255). Was rgb(255, 64, 255).
FAIL colorTest.parseColor("hsl(330, 100%, 62.5%)") should be rgb(255, 63, 159). Was rgb(255, 64, 159).
PASS colorTest.parseColor("hsl(0, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(30, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(60, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(90, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(120, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(150, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(180, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(210, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(240, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(270, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(300, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(330, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(0, 12.5%, 75%)") is "rgb(199, 183, 183)"
PASS colorTest.parseColor("hsl(30, 12.5%, 75%)") is "rgb(199, 191, 183)"
PASS colorTest.parseColor("hsl(60, 12.5%, 75%)") is "rgb(199, 199, 183)"
PASS colorTest.parseColor("hsl(90, 12.5%, 75%)") is "rgb(191, 199, 183)"
PASS colorTest.parseColor("hsl(120, 12.5%, 75%)") is "rgb(183, 199, 183)"
PASS colorTest.parseColor("hsl(150, 12.5%, 75%)") is "rgb(183, 199, 191)"
PASS colorTest.parseColor("hsl(180, 12.5%, 75%)") is "rgb(183, 199, 199)"
PASS colorTest.parseColor("hsl(210, 12.5%, 75%)") is "rgb(183, 191, 199)"
PASS colorTest.parseColor("hsl(240, 12.5%, 75%)") is "rgb(183, 183, 199)"
PASS colorTest.parseColor("hsl(270, 12.5%, 75%)") is "rgb(191, 183, 199)"
PASS colorTest.parseColor("hsl(300, 12.5%, 75%)") is "rgb(199, 183, 199)"
PASS colorTest.parseColor("hsl(330, 12.5%, 75%)") is "rgb(199, 183, 191)"
PASS colorTest.parseColor("hsl(0, 25%, 75%)") is "rgb(207, 175, 175)"
PASS colorTest.parseColor("hsl(30, 25%, 75%)") is "rgb(207, 191, 175)"
PASS colorTest.parseColor("hsl(60, 25%, 75%)") is "rgb(207, 207, 175)"
PASS colorTest.parseColor("hsl(90, 25%, 75%)") is "rgb(191, 207, 175)"
PASS colorTest.parseColor("hsl(120, 25%, 75%)") is "rgb(175, 207, 175)"
PASS colorTest.parseColor("hsl(150, 25%, 75%)") is "rgb(175, 207, 191)"
PASS colorTest.parseColor("hsl(180, 25%, 75%)") is "rgb(175, 207, 207)"
PASS colorTest.parseColor("hsl(210, 25%, 75%)") is "rgb(175, 191, 207)"
PASS colorTest.parseColor("hsl(240, 25%, 75%)") is "rgb(175, 175, 207)"
PASS colorTest.parseColor("hsl(270, 25%, 75%)") is "rgb(191, 175, 207)"
PASS colorTest.parseColor("hsl(300, 25%, 75%)") is "rgb(207, 175, 207)"
PASS colorTest.parseColor("hsl(330, 25%, 75%)") is "rgb(207, 175, 191)"
PASS colorTest.parseColor("hsl(0, 37.5%, 75%)") is "rgb(215, 167, 167)"
PASS colorTest.parseColor("hsl(30, 37.5%, 75%)") is "rgb(215, 191, 167)"
PASS colorTest.parseColor("hsl(60, 37.5%, 75%)") is "rgb(215, 215, 167)"
PASS colorTest.parseColor("hsl(90, 37.5%, 75%)") is "rgb(191, 215, 167)"
PASS colorTest.parseColor("hsl(120, 37.5%, 75%)") is "rgb(167, 215, 167)"
PASS colorTest.parseColor("hsl(150, 37.5%, 75%)") is "rgb(167, 215, 191)"
PASS colorTest.parseColor("hsl(180, 37.5%, 75%)") is "rgb(167, 215, 215)"
PASS colorTest.parseColor("hsl(210, 37.5%, 75%)") is "rgb(167, 191, 215)"
PASS colorTest.parseColor("hsl(240, 37.5%, 75%)") is "rgb(167, 167, 215)"
PASS colorTest.parseColor("hsl(270, 37.5%, 75%)") is "rgb(191, 167, 215)"
PASS colorTest.parseColor("hsl(300, 37.5%, 75%)") is "rgb(215, 167, 215)"
PASS colorTest.parseColor("hsl(330, 37.5%, 75%)") is "rgb(215, 167, 191)"
PASS colorTest.parseColor("hsl(0, 50%, 75%)") is "rgb(223, 159, 159)"
PASS colorTest.parseColor("hsl(30, 50%, 75%)") is "rgb(223, 191, 159)"
PASS colorTest.parseColor("hsl(60, 50%, 75%)") is "rgb(223, 223, 159)"
PASS colorTest.parseColor("hsl(90, 50%, 75%)") is "rgb(191, 223, 159)"
PASS colorTest.parseColor("hsl(120, 50%, 75%)") is "rgb(159, 223, 159)"
PASS colorTest.parseColor("hsl(150, 50%, 75%)") is "rgb(159, 223, 191)"
PASS colorTest.parseColor("hsl(180, 50%, 75%)") is "rgb(159, 223, 223)"
PASS colorTest.parseColor("hsl(210, 50%, 75%)") is "rgb(159, 191, 223)"
PASS colorTest.parseColor("hsl(240, 50%, 75%)") is "rgb(159, 159, 223)"
PASS colorTest.parseColor("hsl(270, 50%, 75%)") is "rgb(191, 159, 223)"
PASS colorTest.parseColor("hsl(300, 50%, 75%)") is "rgb(223, 159, 223)"
PASS colorTest.parseColor("hsl(330, 50%, 75%)") is "rgb(223, 159, 191)"
PASS colorTest.parseColor("hsl(0, 62.5%, 75%)") is "rgb(231, 151, 151)"
PASS colorTest.parseColor("hsl(30, 62.5%, 75%)") is "rgb(231, 191, 151)"
PASS colorTest.parseColor("hsl(60, 62.5%, 75%)") is "rgb(231, 231, 151)"
PASS colorTest.parseColor("hsl(90, 62.5%, 75%)") is "rgb(191, 231, 151)"
PASS colorTest.parseColor("hsl(120, 62.5%, 75%)") is "rgb(151, 231, 151)"
PASS colorTest.parseColor("hsl(150, 62.5%, 75%)") is "rgb(151, 231, 191)"
PASS colorTest.parseColor("hsl(180, 62.5%, 75%)") is "rgb(151, 231, 231)"
PASS colorTest.parseColor("hsl(210, 62.5%, 75%)") is "rgb(151, 191, 231)"
PASS colorTest.parseColor("hsl(240, 62.5%, 75%)") is "rgb(151, 151, 231)"
PASS colorTest.parseColor("hsl(270, 62.5%, 75%)") is "rgb(191, 151, 231)"
PASS colorTest.parseColor("hsl(300, 62.5%, 75%)") is "rgb(231, 151, 231)"
PASS colorTest.parseColor("hsl(330, 62.5%, 75%)") is "rgb(231, 151, 191)"
PASS colorTest.parseColor("hsl(0, 75%, 75%)") is "rgb(239, 143, 143)"
PASS colorTest.parseColor("hsl(30, 75%, 75%)") is "rgb(239, 191, 143)"
PASS colorTest.parseColor("hsl(60, 75%, 75%)") is "rgb(239, 239, 143)"
PASS colorTest.parseColor("hsl(90, 75%, 75%)") is "rgb(191, 239, 143)"
PASS colorTest.parseColor("hsl(120, 75%, 75%)") is "rgb(143, 239, 143)"
PASS colorTest.parseColor("hsl(150, 75%, 75%)") is "rgb(143, 239, 191)"
PASS colorTest.parseColor("hsl(180, 75%, 75%)") is "rgb(143, 239, 239)"
PASS colorTest.parseColor("hsl(210, 75%, 75%)") is "rgb(143, 191, 239)"
PASS colorTest.parseColor("hsl(240, 75%, 75%)") is "rgb(143, 143, 239)"
PASS colorTest.parseColor("hsl(270, 75%, 75%)") is "rgb(191, 143, 239)"
PASS colorTest.parseColor("hsl(300, 75%, 75%)") is "rgb(239, 143, 239)"
PASS colorTest.parseColor("hsl(330, 75%, 75%)") is "rgb(239, 143, 191)"
PASS colorTest.parseColor("hsl(0, 87.5%, 75%)") is "rgb(247, 135, 135)"
PASS colorTest.parseColor("hsl(30, 87.5%, 75%)") is "rgb(247, 191, 135)"
PASS colorTest.parseColor("hsl(60, 87.5%, 75%)") is "rgb(247, 247, 135)"
PASS colorTest.parseColor("hsl(90, 87.5%, 75%)") is "rgb(191, 247, 135)"
PASS colorTest.parseColor("hsl(120, 87.5%, 75%)") is "rgb(135, 247, 135)"
PASS colorTest.parseColor("hsl(150, 87.5%, 75%)") is "rgb(135, 247, 191)"
PASS colorTest.parseColor("hsl(180, 87.5%, 75%)") is "rgb(135, 247, 247)"
PASS colorTest.parseColor("hsl(210, 87.5%, 75%)") is "rgb(135, 191, 247)"
PASS colorTest.parseColor("hsl(240, 87.5%, 75%)") is "rgb(135, 135, 247)"
PASS colorTest.parseColor("hsl(270, 87.5%, 75%)") is "rgb(191, 135, 247)"
PASS colorTest.parseColor("hsl(300, 87.5%, 75%)") is "rgb(247, 135, 247)"
PASS colorTest.parseColor("hsl(330, 87.5%, 75%)") is "rgb(247, 135, 191)"
FAIL colorTest.parseColor("hsl(0, 100%, 75%)") should be rgb(255, 127, 127). Was rgb(255, 128, 128).
FAIL colorTest.parseColor("hsl(30, 100%, 75%)") should be rgb(255, 191, 127). Was rgb(255, 191, 128).
FAIL colorTest.parseColor("hsl(60, 100%, 75%)") should be rgb(255, 255, 127). Was rgb(255, 255, 128).
FAIL colorTest.parseColor("hsl(90, 100%, 75%)") should be rgb(191, 255, 127). Was rgb(191, 255, 128).
FAIL colorTest.parseColor("hsl(120, 100%, 75%)") should be rgb(127, 255, 127). Was rgb(128, 255, 128).
FAIL colorTest.parseColor("hsl(150, 100%, 75%)") should be rgb(127, 255, 191). Was rgb(128, 255, 191).
FAIL colorTest.parseColor("hsl(180, 100%, 75%)") should be rgb(127, 255, 255). Was rgb(128, 255, 255).
FAIL colorTest.parseColor("hsl(210, 100%, 75%)") should be rgb(127, 191, 255). Was rgb(128, 191, 255).
FAIL colorTest.parseColor("hsl(240, 100%, 75%)") should be rgb(127, 127, 255). Was rgb(128, 128, 255).
FAIL colorTest.parseColor("hsl(270, 100%, 75%)") should be rgb(191, 127, 255). Was rgb(191, 128, 255).
FAIL colorTest.parseColor("hsl(300, 100%, 75%)") should be rgb(255, 127, 255). Was rgb(255, 128, 255).
FAIL colorTest.parseColor("hsl(330, 100%, 75%)") should be rgb(255, 127, 191). Was rgb(255, 128, 191).
PASS colorTest.parseColor("hsl(0, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(30, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(60, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(90, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(120, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(150, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(180, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(210, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(240, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(270, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(300, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(330, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(0, 12.5%, 87.5%)") is "rgb(227, 219, 219)"
PASS colorTest.parseColor("hsl(30, 12.5%, 87.5%)") is "rgb(227, 223, 219)"
PASS colorTest.parseColor("hsl(60, 12.5%, 87.5%)") is "rgb(227, 227, 219)"
PASS colorTest.parseColor("hsl(90, 12.5%, 87.5%)") is "rgb(223, 227, 219)"
PASS colorTest.parseColor("hsl(120, 12.5%, 87.5%)") is "rgb(219, 227, 219)"
PASS colorTest.parseColor("hsl(150, 12.5%, 87.5%)") is "rgb(219, 227, 223)"
PASS colorTest.parseColor("hsl(180, 12.5%, 87.5%)") is "rgb(219, 227, 227)"
PASS colorTest.parseColor("hsl(210, 12.5%, 87.5%)") is "rgb(219, 223, 227)"
PASS colorTest.parseColor("hsl(240, 12.5%, 87.5%)") is "rgb(219, 219, 227)"
PASS colorTest.parseColor("hsl(270, 12.5%, 87.5%)") is "rgb(223, 219, 227)"
PASS colorTest.parseColor("hsl(300, 12.5%, 87.5%)") is "rgb(227, 219, 227)"
PASS colorTest.parseColor("hsl(330, 12.5%, 87.5%)") is "rgb(227, 219, 223)"
PASS colorTest.parseColor("hsl(0, 25%, 87.5%)") is "rgb(231, 215, 215)"
PASS colorTest.parseColor("hsl(30, 25%, 87.5%)") is "rgb(231, 223, 215)"
PASS colorTest.parseColor("hsl(60, 25%, 87.5%)") is "rgb(231, 231, 215)"
PASS colorTest.parseColor("hsl(90, 25%, 87.5%)") is "rgb(223, 231, 215)"
PASS colorTest.parseColor("hsl(120, 25%, 87.5%)") is "rgb(215, 231, 215)"
PASS colorTest.parseColor("hsl(150, 25%, 87.5%)") is "rgb(215, 231, 223)"
PASS colorTest.parseColor("hsl(180, 25%, 87.5%)") is "rgb(215, 231, 231)"
PASS colorTest.parseColor("hsl(210, 25%, 87.5%)") is "rgb(215, 223, 231)"
PASS colorTest.parseColor("hsl(240, 25%, 87.5%)") is "rgb(215, 215, 231)"
PASS colorTest.parseColor("hsl(270, 25%, 87.5%)") is "rgb(223, 215, 231)"
PASS colorTest.parseColor("hsl(300, 25%, 87.5%)") is "rgb(231, 215, 231)"
PASS colorTest.parseColor("hsl(330, 25%, 87.5%)") is "rgb(231, 215, 223)"
PASS colorTest.parseColor("hsl(0, 37.5%, 87.5%)") is "rgb(235, 211, 211)"
PASS colorTest.parseColor("hsl(30, 37.5%, 87.5%)") is "rgb(235, 223, 211)"
PASS colorTest.parseColor("hsl(60, 37.5%, 87.5%)") is "rgb(235, 235, 211)"
PASS colorTest.parseColor("hsl(90, 37.5%, 87.5%)") is "rgb(223, 235, 211)"
PASS colorTest.parseColor("hsl(120, 37.5%, 87.5%)") is "rgb(211, 235, 211)"
PASS colorTest.parseColor("hsl(150, 37.5%, 87.5%)") is "rgb(211, 235, 223)"
PASS colorTest.parseColor("hsl(180, 37.5%, 87.5%)") is "rgb(211, 235, 235)"
PASS colorTest.parseColor("hsl(210, 37.5%, 87.5%)") is "rgb(211, 223, 235)"
PASS colorTest.parseColor("hsl(240, 37.5%, 87.5%)") is "rgb(211, 211, 235)"
PASS colorTest.parseColor("hsl(270, 37.5%, 87.5%)") is "rgb(223, 211, 235)"
PASS colorTest.parseColor("hsl(300, 37.5%, 87.5%)") is "rgb(235, 211, 235)"
PASS colorTest.parseColor("hsl(330, 37.5%, 87.5%)") is "rgb(235, 211, 223)"
PASS colorTest.parseColor("hsl(0, 50%, 87.5%)") is "rgb(239, 207, 207)"
PASS colorTest.parseColor("hsl(30, 50%, 87.5%)") is "rgb(239, 223, 207)"
PASS colorTest.parseColor("hsl(60, 50%, 87.5%)") is "rgb(239, 239, 207)"
PASS colorTest.parseColor("hsl(90, 50%, 87.5%)") is "rgb(223, 239, 207)"
PASS colorTest.parseColor("hsl(120, 50%, 87.5%)") is "rgb(207, 239, 207)"
PASS colorTest.parseColor("hsl(150, 50%, 87.5%)") is "rgb(207, 239, 223)"
PASS colorTest.parseColor("hsl(180, 50%, 87.5%)") is "rgb(207, 239, 239)"
PASS colorTest.parseColor("hsl(210, 50%, 87.5%)") is "rgb(207, 223, 239)"
PASS colorTest.parseColor("hsl(240, 50%, 87.5%)") is "rgb(207, 207, 239)"
PASS colorTest.parseColor("hsl(270, 50%, 87.5%)") is "rgb(223, 207, 239)"
PASS colorTest.parseColor("hsl(300, 50%, 87.5%)") is "rgb(239, 207, 239)"
PASS colorTest.parseColor("hsl(330, 50%, 87.5%)") is "rgb(239, 207, 223)"
PASS colorTest.parseColor("hsl(0, 62.5%, 87.5%)") is "rgb(243, 203, 203)"
PASS colorTest.parseColor("hsl(30, 62.5%, 87.5%)") is "rgb(243, 223, 203)"
PASS colorTest.parseColor("hsl(60, 62.5%, 87.5%)") is "rgb(243, 243, 203)"
PASS colorTest.parseColor("hsl(90, 62.5%, 87.5%)") is "rgb(223, 243, 203)"
PASS colorTest.parseColor("hsl(120, 62.5%, 87.5%)") is "rgb(203, 243, 203)"
PASS colorTest.parseColor("hsl(150, 62.5%, 87.5%)") is "rgb(203, 243, 223)"
PASS colorTest.parseColor("hsl(180, 62.5%, 87.5%)") is "rgb(203, 243, 243)"
PASS colorTest.parseColor("hsl(210, 62.5%, 87.5%)") is "rgb(203, 223, 243)"
PASS colorTest.parseColor("hsl(240, 62.5%, 87.5%)") is "rgb(203, 203, 243)"
PASS colorTest.parseColor("hsl(270, 62.5%, 87.5%)") is "rgb(223, 203, 243)"
PASS colorTest.parseColor("hsl(300, 62.5%, 87.5%)") is "rgb(243, 203, 243)"
PASS colorTest.parseColor("hsl(330, 62.5%, 87.5%)") is "rgb(243, 203, 223)"
PASS colorTest.parseColor("hsl(0, 75%, 87.5%)") is "rgb(247, 199, 199)"
PASS colorTest.parseColor("hsl(30, 75%, 87.5%)") is "rgb(247, 223, 199)"
PASS colorTest.parseColor("hsl(60, 75%, 87.5%)") is "rgb(247, 247, 199)"
PASS colorTest.parseColor("hsl(90, 75%, 87.5%)") is "rgb(223, 247, 199)"
PASS colorTest.parseColor("hsl(120, 75%, 87.5%)") is "rgb(199, 247, 199)"
PASS colorTest.parseColor("hsl(150, 75%, 87.5%)") is "rgb(199, 247, 223)"
PASS colorTest.parseColor("hsl(180, 75%, 87.5%)") is "rgb(199, 247, 247)"
PASS colorTest.parseColor("hsl(210, 75%, 87.5%)") is "rgb(199, 223, 247)"
PASS colorTest.parseColor("hsl(240, 75%, 87.5%)") is "rgb(199, 199, 247)"
PASS colorTest.parseColor("hsl(270, 75%, 87.5%)") is "rgb(223, 199, 247)"
PASS colorTest.parseColor("hsl(300, 75%, 87.5%)") is "rgb(247, 199, 247)"
PASS colorTest.parseColor("hsl(330, 75%, 87.5%)") is "rgb(247, 199, 223)"
PASS colorTest.parseColor("hsl(0, 87.5%, 87.5%)") is "rgb(251, 195, 195)"
PASS colorTest.parseColor("hsl(30, 87.5%, 87.5%)") is "rgb(251, 223, 195)"
PASS colorTest.parseColor("hsl(60, 87.5%, 87.5%)") is "rgb(251, 251, 195)"
PASS colorTest.parseColor("hsl(90, 87.5%, 87.5%)") is "rgb(223, 251, 195)"
PASS colorTest.parseColor("hsl(120, 87.5%, 87.5%)") is "rgb(195, 251, 195)"
PASS colorTest.parseColor("hsl(150, 87.5%, 87.5%)") is "rgb(195, 251, 223)"
PASS colorTest.parseColor("hsl(180, 87.5%, 87.5%)") is "rgb(195, 251, 251)"
PASS colorTest.parseColor("hsl(210, 87.5%, 87.5%)") is "rgb(195, 223, 251)"
PASS colorTest.parseColor("hsl(240, 87.5%, 87.5%)") is "rgb(195, 195, 251)"
PASS colorTest.parseColor("hsl(270, 87.5%, 87.5%)") is "rgb(223, 195, 251)"
PASS colorTest.parseColor("hsl(300, 87.5%, 87.5%)") is "rgb(251, 195, 251)"
PASS colorTest.parseColor("hsl(330, 87.5%, 87.5%)") is "rgb(251, 195, 223)"
PASS colorTest.parseColor("hsl(0, 100%, 87.5%)") is "rgb(255, 191, 191)"
PASS colorTest.parseColor("hsl(30, 100%, 87.5%)") is "rgb(255, 223, 191)"
PASS colorTest.parseColor("hsl(60, 100%, 87.5%)") is "rgb(255, 255, 191)"
PASS colorTest.parseColor("hsl(90, 100%, 87.5%)") is "rgb(223, 255, 191)"
PASS colorTest.parseColor("hsl(120, 100%, 87.5%)") is "rgb(191, 255, 191)"
PASS colorTest.parseColor("hsl(150, 100%, 87.5%)") is "rgb(191, 255, 223)"
PASS colorTest.parseColor("hsl(180, 100%, 87.5%)") is "rgb(191, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 87.5%)") is "rgb(191, 223, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 87.5%)") is "rgb(191, 191, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 87.5%)") is "rgb(223, 191, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 87.5%)") is "rgb(255, 191, 255)"
PASS colorTest.parseColor("hsl(330, 100%, 87.5%)") is "rgb(255, 191, 223)"
PASS colorTest.parseColor("hsl(0, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 100%, 100%)") is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE
